rename form fields file
